requirements-test.txt
uWSGI==2.0.18
wheel==0.34.2
psycopg2-binary==2.8.5
marshmallow==3.6.1
Flask-Testing==0.8.0
Flask-Script==2.0.6
flask-restx==0.2.0
Flask-Migrate==2.5.3
boto3==1.13.20
coverage==4.5.4
flake8==3.8.3
flake8-docstrings==1.5.0
flake8-import-order==0.18.1
wait-for-it==1.1.2